requirements.txt
flask==2.3.3
openai==1.6.1
pandas==2.0.3
numpy==1.24.4
python-dotenv==1.0.0
gunicorn==21.2.0
markupsafe==2.1.3
werkzeug==2.3.7
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7 